@@ -16,54 +16,54 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^7.2.12",
-    "@angular/core": "^7.2.12",
-    "@angular/compiler": "^7.2.12",
-    "@angular/animations": "^7.2.12",
-    "@angular/forms": "^7.2.12",
-    "@angular/http": "^7.2.12",
-    "@angular/platform-browser": "^7.2.12",
-    "@angular/platform-browser-dynamic": "^7.2.12",
-    "@angular/router": "^7.2.12",
-    "core-js": "^2.6.3",
-    "rxjs": "^6.4.0",
-    "zone.js": "^0.9.0",
+    "@angular/common": "^8.1.1",
+    "@angular/core": "^8.1.1",
+    "@angular/compiler": "^8.1.1",
+    "@angular/animations": "^8.1.1",
+    "@angular/forms": "^8.1.1",
+    "@angular/http": "^7.2.15",
+    "@angular/platform-browser": "^8.1.1",
+    "@angular/platform-browser-dynamic": "^8.1.1",
+    "@angular/router": "^8.1.1",
+    "core-js": "^3.1.4",
+    "rxjs": "^6.5.2",
+    "zone.js": "^0.9.1",
     "toastr": "^2.1.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.8",
-    "@angular-devkit/build-ng-packagr": "~0.13.8",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler-cli": "^7.2.12",
-    "@angular/language-service": "^7.2.12",
-    "@ng-bootstrap/ng-bootstrap": "^4.1.1",
-    "@types/jasmine": "~3.3.12",
+    "@angular-devkit/build-angular": "~0.801.1",
+    "@angular-devkit/build-ng-packagr": "~0.801.1",
+    "@angular/cli": "~8.1.1",
+    "@angular/compiler-cli": "^8.1.1",
+    "@angular/language-service": "^8.1.1",
+    "@ng-bootstrap/ng-bootstrap": "^5.0.0",
+    "@types/jasmine": "~3.3.13",
     "@types/jasminewd2": "~2.0.6",
-    "@types/lodash": "^4.14.123",
-    "@types/node": "~11.13.0",
+    "@types/lodash": "^4.14.136",
+    "@types/node": "~12.6.3",
     "@types/toastr": "^2.1.37",
-    "codelyzer": "~5.0.0",
-    "gulp": "^4.0.0",
+    "codelyzer": "~5.1.0",
+    "gulp": "^4.0.2",
     "gulp-bump": "^3.1.3",
     "gulp-clean": "^0.4.0",
     "gulp-git": "^2.9.0",
     "gulp-help-doc": "^1.1.1",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.1",
-    "karma-chrome-launcher": "~2.2.0",
+    "karma": "~4.2.0",
+    "karma-chrome-launcher": "~3.0.0",
     "karma-coverage-istanbul-reporter": "~2.0.5",
     "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "ng-packagr": "^5.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "ng-packagr": "^5.3.0",
     "protractor": "~5.4.2",
     "require-dir": "^1.2.0",
     "run-sequence": "^2.2.1",
-    "ts-node": "~8.0.3",
-    "tsickle": ">=0.34.3",
-    "tslib": "^1.9.3",
-    "tslint": "~5.15.0",
-    "typescript": "^3.1.1",
-    "yargs": "^13.2.2"
+    "ts-node": "~8.3.0",
+    "tsickle": ">=0.35.0",
+    "tslib": "^1.10.0",
+    "tslint": "~5.18.0",
+    "typescript": "^3.4.5",
+    "yargs": "^13.3.0"
   }
 }
