@@ -10,9 +10,9 @@
     "javascript"
   ],
   "peerDependencies": {
-    "@angular/common": "^7.0.0",
-    "@angular/core": "^7.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^3.0.0",
+    "@angular/common": "^8.1.1",
+    "@angular/core": "^8.1.1",
+    "@ng-bootstrap/ng-bootstrap": "^5.0.0",
     "toastr": "^2.1.4"
   }
 }
